+++ samod/tests/tincans.rs
@@ -1,5 +1,5 @@
 #![allow(dead_code)]
-use std::{pin::Pin, sync::Arc};
+use std::{convert::Infallible, pin::Pin, sync::Arc};
 
 use futures::{FutureExt, Sink, SinkExt, Stream, StreamExt, select};
 use rand::Rng;
@@ -96,6 +96,8 @@ struct CancellableDialer {
 }
 
 impl Dialer for CancellableDialer {
+    type Error = Infallible;
+
     fn url(&self) -> Url {
         self.url.clone()
     }
@@ -105,7 +107,7 @@ impl Dialer for CancellableDialer {
     ) -> Pin<
         Box<
             dyn std::future::Future<
-                    Output = Result<Transport, Box<dyn std::error::Error + Send + Sync + 'static>>,
+                    Output = Result<Transport, samod::DialError<Self::Error>>,
                 > + Send,
         >,
     > {
@@ -115,7 +117,7 @@ impl Dialer for CancellableDialer {
         Box::pin(async move {
             acceptor
                 .accept(Transport::new(acceptor_side.recv, acceptor_side.send))
-                .map_err(|e| Box::new(e) as Box<dyn std::error::Error + Send + Sync + 'static>)?;
+                .map_err(samod::DialError::transient)?;
 
             Ok(Transport::new(dialer_side.recv, dialer_side.send))
         })